Add scan feature plan metadata helpers
